Few Changes
Added Instructions, spell-checked some files, added a few files.
